webpack-config > Fix broken production build
\### Rationale
Given the way that the configuration is generated with some static
values versus inline functions the previous change to only minimize on
production builds was set to off -- always.  Fixed this to use the
condition that will in-fact detect if it is currently a production
build or not.  Ideal state is to not minimize on development builds so
that a developer can step through the code via a debugger.

Without this fix, the production build was maintaining all the comments
and extra jazz from webpack.