webpack-config > Ensure `node_modules` will not bundle
\### Rationale
This is more of a future-proofing step to ensure that later on no
runtime dependencies are bundled into the core functionality of this
library but instead are always considered an external reference.

There are no current dependencies for this library at this time but
even so this adds some rules for the future.