lint-config > Change config for better `eslint-mdx`
\### Rationale
Based on the issue resolution I received, the configuration was not
how I expected for the `mdx` plugin.  It ends up the plugin does not
verify any of the files it receives but expects eslint to provide it
the correct extension files.

With version 1.11.0 performance is still poor as the program hangs and
finishes at 25 seconds or so.