lint-config > Upgrade `eslint-plugin-mdx`
\### Rationale
With the upgrade to the latest plugin version, performance is much
better with the new config. No `TypeError` as it was providing before.

Performance actually resolves in a reasonable amount of time. A
performance check resulted in a 10 second runtime and it actually quit
the program instead of the ending hang.