lint-errors > Fixed TS codeblock parsing!
\### Rationale
The `@typescript-eslint/parser` is flawed to require `tsconfig.json` to
ensure certain rule conditions related to type checking are enforced.
Furthermore the developers recognize that Markdown codeblocks are not
supported via the virtual filepath they use inside of the eslint
framework. I found a workaround by excluding from the traditional ts
override configuration into its own block. I systematically ignored
the specific rules that require `tsconfig` to evaluate files on disk
and now it works and still parses Typescript!

Also extracted a few of the duplicate configurations out to constants
at the top of the config file. Thankfully it is `js`